fix(ingestion/airflow-plugin): fixed review comments
