+++ metadata-ingestion-modules/airflow-plugin/src/datahub_airflow_plugin/_datahub_ol_adapter.py
@@ -14,15 +14,10 @@ OL_SCHEME_TWEAKS = {
 
 
 def translate_ol_to_datahub_urn(ol_uri: OpenLineageDataset) -> str:
-    if not hasattr(ol_uri, "namespace") or not ol_uri.namespace:
-        return builder.make_dataset_urn(
-            platform=getattr(ol_uri, "platform", "unknown"), name=ol_uri.name
-        )
-    else:
-        namespace = ol_uri.namespace
-        name = ol_uri.name
+    namespace = ol_uri.namespace
+    name = ol_uri.name
 
-        scheme, *rest = namespace.split("://", maxsplit=1)
+    scheme, *rest = namespace.split("://", maxsplit=1)
 
-        platform = OL_SCHEME_TWEAKS.get(scheme, scheme)
-        return builder.make_dataset_urn(platform=platform, name=name)
+    platform = OL_SCHEME_TWEAKS.get(scheme, scheme)
+    return builder.make_dataset_urn(platform=platform, name=name)
