+++ metadata-ingestion-modules/airflow-plugin/tests/integration/test_plugin.py
@@ -113,7 +113,6 @@ def _wait_for_dag_finish(
 def _run_airflow(
     tmp_path: pathlib.Path,
     dags_folder: pathlib.Path,
-    plugins_folder: pathlib.Path,
     is_v1: bool,
 ) -> Iterator[AirflowInstance]:
     airflow_home = tmp_path / "airflow_home"
@@ -136,7 +135,6 @@ def _run_airflow(
         # Point airflow to the DAGs folder.
         "AIRFLOW__CORE__LOAD_EXAMPLES": "False",
         "AIRFLOW__CORE__DAGS_FOLDER": str(dags_folder),
-        "AIRFLOW__CORE__PLUGINS_FOLDER": str(plugins_folder),
         "AIRFLOW__CORE__DAGS_ARE_PAUSED_AT_CREATION": "False",
         # Have the Airflow API use username/password authentication.
         "AIRFLOW__API__AUTH_BACKEND": "airflow.api.auth.backend.basic_auth",
@@ -346,7 +344,7 @@ def test_airflow_plugin(
     dag_id = test_case.dag_id
 
     with _run_airflow(
-        tmp_path, dags_folder=DAGS_FOLDER, plugins_folder=PLUGINS_FOLDER, is_v1=is_v1
+        tmp_path, dags_folder=DAGS_FOLDER, is_v1=is_v1
     ) as airflow_instance:
         print(f"Running DAG {dag_id}...")
         subprocess.check_call(
@@ -428,7 +426,6 @@ if __name__ == "__main__":
     with _run_airflow(
         tmp_path=pathlib.Path(tempfile.mkdtemp("airflow-plugin-test")),
         dags_folder=DAGS_FOLDER,
-        plugins_folder=PLUGINS_FOLDER,
         is_v1=not HAS_AIRFLOW_LISTENER_API,
     ) as airflow_instance:
         # input("Press enter to exit...")
